moved stuff around and used modal instead
