+++ apolloschurchapp/schema.graphql
@@ -581,7 +581,6 @@ type Query {
   campaigns: ContentItemsConnection
   userFeed(first: Int, after: String): ContentItemsConnection
   personaFeed(first: Int, after: String): ContentItemsConnection
-  suggestedFollows: [Person]
   currentUser: AuthenticatedUser
   userExists(identity: String): USER_AUTH_STATUS
   liveStream: LiveStream @deprecated(reason: "Use liveStreams, there may be multiple.")
@@ -804,3 +803,4 @@ type WeekendContentItem implements ContentItem & Node & ContentNode & Card & Vid
   features: [Feature] @deprecated(reason: "Use featureFeed")
   featureFeed: FeatureFeed
 }
+
