--- apolloschurchapp/src/tabs/home/Home.js
@@ -1,76 +1,0 @@
-import React from 'react';
-import { Image } from 'react-native';
-import { SafeAreaView } from 'react-native-safe-area-context';
-import gql from 'graphql-tag';
-import { Query } from '@apollo/client/react/components';
-import { useNavigation } from '@react-navigation/native';
-
-import { styled, BackgroundView } from '@apollosproject/ui-kit';
-import {
-  FeaturesFeedConnected,
-  FEATURE_FEED_ACTION_MAP,
-  RockAuthedWebBrowser,
-} from '@apollosproject/ui-connected';
-
-import HomeSearchButton from './HomeSearchButton';
-
-const LogoTitle = styled(({ theme }) => ({
-  height: theme.sizing.baseUnit,
-  margin: theme.sizing.baseUnit,
-  alignSelf: 'center',
-  resizeMode: 'contain',
-}))(Image);
-
-function handleOnPress({ action, ...props }) {
-  if (FEATURE_FEED_ACTION_MAP[action]) {
-    FEATURE_FEED_ACTION_MAP[action]({ action, ...props });
-  }
-  // If you add additional actions, you can handle them here.
-  // Or add them to the FEATURE_FEED_ACTION_MAP, with the syntax
-  // { [ActionName]: function({ relatedNode, action, ...FeatureFeedConnectedProps}) }
-}
-
-// getHomeFeed uses the HOME_FEATURES in the config.yml
-// You can also hardcode an ID if you are confident it will never change
-// Or use some other strategy to get a FeatureFeed.id
-export const GET_HOME_FEED = gql`
-  query getHomeFeatureFeed {
-    homeFeedFeatures {
-      id
-    }
-  }
-`;
-
-const Home = () => {
-  const navigation = useNavigation();
-  return (
-    <RockAuthedWebBrowser>
-      {(openUrl) => (
-        <BackgroundView>
-          <SafeAreaView edges={['top', 'left', 'right']}>
-            <Query query={GET_HOME_FEED} fetchPolicy="cache-and-network">
-              {({ data }) => (
-                <FeaturesFeedConnected
-                  openUrl={openUrl}
-                  navigation={navigation}
-                  featureFeedId={data?.homeFeedFeatures?.id}
-                  onPressActionItem={handleOnPress}
-                  ListHeaderComponent={
-                    <>
-                      <LogoTitle source={require('./wordmark.png')} />
-                      <HomeSearchButton
-                        onPress={() => navigation.navigate('Search')}
-                      />
-                    </>
-                  }
-                />
-              )}
-            </Query>
-          </SafeAreaView>
-        </BackgroundView>
-      )}
-    </RockAuthedWebBrowser>
-  );
-};
-
-export default Home;
