+++ apolloschurchapp/src/ui/Search/index.js
@@ -1,0 +1,5 @@
+import Search from './Search';
+import SearchButton from './SearchButton';
+
+export { SearchButton };
+export default Search;
